--- a/pdfminer/__init__.py
+++ b/pdfminer/__init__.py
@@ -10,7 +10,7 @@ It includes a PDF converter that can transform PDF files into other text
 formats (such as HTML). It has an extensible PDF parser that can be used for
 other purposes instead of text analysis.
 """
-__version__ = '20170720'
+__version__ = '20181108'
 
 if __name__ == '__main__':
     print(__version__)