@mikkkee 2019-10-23
Hi @pietermarsman  , thanks, I just opened https://github.com/pdfminer/pdfminer.six/pull/315